fix(tokenizer): Allow multi-line comments that aren't followed by a space
Currently the start of a multi-line comment is only parsed as such if
the tokenizer is in one of the junk states; that is, if it is followed
by a space or newline. Starts of multi-line comments that occur at the
beginning of the source text or immediately after the end of a token
are misparsed as a division punctuator followed by a multiplication
punctuator. This change fixes this.