refactor: use better `Value` and sub-enum names
This should make a lot of our code easier to understand by displaying
the number of bits in the name of our `I56` type used to niche `Number`
and `BigInt`. Furthermore, I made the niches start with the original
type in the `Value` enum and their subtypes (`Number` and `BigInt`)
which should also help with readability.