chore(vm): Refactor `Vm::execute` to allow for try-catch (#181)
The current implementation of `Vm::execute` uses the `?` operator to
propagate JS exceptions to the caller. This currently only works
because we don't implement try-catch, since `?` does not allow
catching an `Err` result value within the current function.

This patch adds an `execute_instruction` method, which handles the
execution of a single instruction, which lets the `execute` method
handle its non-successful result value.

However, this would not work with `return` statements, since they are
implemented by returning a value from `execute`. To fix this, we add a
`ContinuationKind` enum, which we use as the successful return type of
`execute_instruction`, to be able to tell apart what the spec calls
normal vs return completions.