fix(ecmascript): Make sure built-in functions have `name` and `length`
When built-in functions are created, the `initial_name` and `length`
fields of `BuiltinFunctionHeapData` are populated. This means that, if
there is no backing object when the function is created, the `name`
and `length` properties will have the value given by those fields.
This is the case even if a backing object needs to be created later
on, since those fields will be added to the backing object.

However, if properties are added to the built-in function as it is
being constructed, then there will be a backing object without the
`name` and `length` properties, so they will appear to not be set.
This is always the case for built-in constructor functions, since they
will always have a `prototype` property.

The long-term fix for this would be to avoid duplicating heap data
fields in the backing object somehow. However, in the meantime, this
patch makes sure that all built-in functions that are created with a
backing object have the `name` and `length` properties in it.