feat(ecmascript): Allow object creation with a different prototype
Per the spec, OrdinaryCreateFromConstructor should get the `prototype`
property of the passed constructor, and if that is an object, it
should create a new object whose prototype is that property;
otherwise, the prototype would be set to the passed intrinsic default
prototype.

In the current implementation of this method, the `prototype` of the
passed constructor is obtained, but then it is ignored. The given
intrinsic default prototype is used to determine in which heap to
store the object, which gives it the right default prototype, but
it does not check whether the constructor's `prototype` property
matches that default prototype.

It seems like this gets the right result for all of the times where
Nova currently does object construction. However, allowing creating an
object with a prototype different from the heap's default would be
needed for constructing ECMAScript functions.

As part of that, this patch changes `get_prototype_from_constructor`
to return a `JsResult<Option<Object>>`, guaranteeing that it will not
give the caller an object which is the `ProtoIntrinsics`'s
corresponding prototype. It also changes
`ordinary_object_create_with_intrinsics` to take an optional prototype
object aside from the optional `ProtoIntrinsics`.

It also changes `ordinary_create_from_constructor` to use the
prototype obtained from the constructor, as well as to remove the
`_internal_slots_list` parameter. In the spec this parameter is used
to create the object with the right internal slots, but in Nova the
equivalent of this is choosing in which heap the object is created,
which the `ProtoIntrinsics` argument already does.