feat(ecmascript): Implement the `this` keyword
This patch also fixes a bug in which functions created with the
`InstantiateOrdinaryFunctionExpression` bytecode instruction were
created with a lexical this mode, when it should be either strict or
global. The lexical this mode is only used for arrow functions, and it
causes the `this` resolution to be wrong in other functions.