feat(nova_cli): Print uncaught exceptions
This patch prints a string representation of an uncaught exception in
`nova_cli eval` before exiting the process with error code 1.

Since the `ToString()` operation is not guaranteed to succeed, this
instead adds a `Value::string_repr` which never returns `Err`, and
which accounts for symbols as well as objects whose stringification
throws.