fix(small_string): Support null-terminated small strings (#237)
Currently small strings use the zero byte as filler between the end of
the string and the end of the buffer. This makes it impossible to
represent strings ending in `\0` as a small string, even if they fit
the length requirements.

However, since small strings are always UTF-8, it is possible to use
a byte like 0xFF as the filler byte, since valid UTF-8 (or WTF-8)
strings will never contain that byte. This makes it possible to
represent the null character at the end of a small string while
distinguishing it from the end of the string.