fix(ecmascript): Only implicit return in concise arrow functions
If a function ends in a non-return expression, currently Nova will
use that as the return value for the function. In JS this should only
be the case for concise arrow functions – i.e, arrow functions which
have a single expression, rather than a block, as their body – and the
return value for any other functions should be `undefined`.

Aside from arrow functions, one big reason why Nova was returning
implicitly is that the return value of `Vm::execute` was being used
both as the return value for functions, as well as the result of
script evaluation, which is used in script tests. To avoid breaking
them, this function adds the `ExecutionResult` enum, which lets the
return value of `Vm::execute` distinguish between JS returns vs
non-returns.