fix(ecmascript): Fix `Object.is{Extensible,Sealed,Frozen}`
The implementation of `Object.isSealed` and `Object.isFrozen`,
although it did not panic, it just returned the passed object, which
is never a valid behavior per the spec. Although `Object.isExtensible`
was implemented per the spec for objects, for primitives it returned
the passed value rather than `false`. This patch fixes these issues.