fix(ecmascript): Enforce the spec's limits on string length (#287)
According to the spec, the maximum length of a string is 2^53 - 1
UTF-16 code units. Since Nova's strings are UTF-8, and we only compute
their UTF-16 length and indices lazily, we cannot fulfill this
restriction as is. Instead, when a heap string is created, we check
that its UTF-8 length could correspond to a valid UTF-16 length. Then,
when and if we compute the UTF-16 mapping, we panic if that would
result in an invalid UTF-16 length.