refactor(ecmascript): Refactor `PromiseCapability` and `PromiseState`
In the spec, Promise Capability Records contain a promise-like object
as well as its resolve and reject functions. This is needed if you
allow promise subclassing, which makes it so the resolve and reject
functions might not be the ones the spec defines. However, for Nova's
implementation of promises, we're only allowing built-in Promises (as
well as thenables).

If we see a promise capability as giving you both a promise and a way
to resolve it, we can instead implement that as a wrapper over
`Promise` that gives you built-in `resolve` and `reject` methods. The
promise resolving functions could then build *on top of*
`PromiseCapability`, rather than vice versa. This would also mean that
there wouldn't be a heap vector for promise capabilities themselves.

There's one remaining issue: there are up to two times where resolving
functions can be created for a promise. One is when it is created (and
those resolving functions are fed to the construction callback).
Another is when a promise is resolved onto another promise or
thenable (and those resolving functions are passed to its `.then()`
method).

Each pair of promise resolving functions has a shared invalidation
condition, which causes those functions to not work anymore. In the
spec, a function pair is invalidated when either function is called.
Here, we want to represent this in the promise state instead: the pair
created during construction is valid if the promise is pending and
unresolved, the pair created during resolve is valid if the promise is
pending (and resolved).

To track this difference, we add an extra `must_be_unresolved` boolean
to `PromiseCapability`, and we refactor `PromiseState` to add an
`is_resolved` boolean to the `Pending` state. While we're at it, we
also move the fulfill and rejection reactions from `PromiseHeapData`
to the `PromiseState::Pending` state, and the `promise_is_handled`
flag to the rejected state.