feat(ecmascript, cli): Implement promise jobs (#303)
The last piece of the basic Promise implementation, this patch
implements promise jobs, as well as the `HostEnqueuePromiseJob` host
hook that gives them to the embedder to enqueue and later run.

Although the spec makes a difference between various kinds of jobs
(e.g. promise vs timeout vs `FinalizationRegistry` cleanup jobs), and
implementations typically treat them differently for scheduling, the
spec does not require that they are stored in separate queues. We
therefore have a single embedder-facing `Job` type, so jobs of
different kinds might be stored in the same queue. This type wraps an
enum that disambiguates between the different internal job structs,
each with a `run()` implementation.

This patch also adds the `HostPromiseRejectionTracker` hook, meant to
enable hosts to track promises that have rejected without catch
handlers (e.g. for the `unhandledrejection` HTML event).