feat(ecmascript): Implement strict vs. non-strict code distinction
Whenever the spec said to have different behaviors for strict and
non-strict code, we previously sometimes went with the strict and
sometimes with the loose behavior, because we weren't tracking whether
the current execution was strict or not. This patch tracks that.

Ideally, oxc would store whether a function or not is strict code
(i.e. whether it has a `"use strict"` declaration or its lexical
parent was strict code), since it needs to have that information in
order to properly parse it. However, it only exposes whether it has
the `"use strict"` declaration. So we track whether the current code
is strict mode in `ECMAScriptCodeEvaluationState`. This also lets us
use that information for direct `eval`.

We don't currently implement mapped `arguments` objects, which are
needed for non-strict mode function. In order to avoid blocking this
patch on implementing that, or having most non-strict functions crash
(since we currently create an `arguments` object for every function),
this patch uses unmapped `arguments` objects always. Mapped
`arguments` objects are left for later.

This patch also changes `OrdinaryFunctionCreateParams`'s `this_mode`
parameter, which used to be the `ThisMode` enum, into a `lexical_this`
bool, which better reflects the spec's distinction between the
"lexical this" and "non lexical this" states. We now have the
information to translate that into the `ThisMode` enum inside the
`ordinary_function_create` function.

When running code with nova_cli, by default it will be run as strict
mode, unless the `--no-strict` flag is passed. As for test262, most
tests must be run twice, once in strict and once in loose mode, to
verify that the behavior does not change. Since we don't currently
support multiple expectations for a single test, we run those tests in
strict mode, and only run `noStrict` tests in loose mode. We also add
a `--prefer-loose-mode` flag to instead only run `onlyStrict` tests in
strict mode.