feat(ecmascript): Support strict mode in bindings and references
PR #319 implemented the distinction between strict and non-strict mode
code, but it did not update every case where the behavior depended on
the strictness of the code. In particular, it did not update the
behavior of bindings and references, which depends on strictness.

As it turns out, the test262 harness was assigning to an uninitialized
identifier as a way of setting a the `$262` property on the global
object, but this throws an error in strict mode, so this patch also
fixes that.