refactor: Allow storing VM and execution state in GC'd types (#363)
* refactor: Allow storing VM and execution state in GC'd types

In order to allow `await` and `yield` to work, the execution state of
a function at the time of evaluating an `await` or `yield` expression
must be stored as part of one of the heap vectors, so it can later be
resumed.

This means that `Vm`, `Executable` and `ExecutionContext` must
implement `HeapMarkAndSweep`. This patch adds those implementations to
those types, as well as many other internal types that did not already
support GC tracing.

All such types must also implement `Send` and `Sync`, since the GC is
meant to become multi-threaded in the future. This was not the case
for `Executable`, since it includes static references into the oxc
allocator, which is single-threaded. However, it would be fine to
unsafely send those references across threads as long as they're only
dereferenced in the main thread. So we add a `SendableRef` type to
make this possible.

* Fix

* Don't forget to track `ECMAScriptCodeEvaluationState::source_code`