refactor: Refactor promise reactions in preparation for `await` (#362)
In the spec, there are only two types of promise reaction handlers:
callback handlers and empty handlers. The internal behavior of `await`
and of other spec-internal machinery that depends on promise reactions
simply creates JS functions backed by a spec-internal abstract closure
and uses them as callback handlers.

These JS functions, however, should not be observable from JS code,
and in fact they should only ever be handled by the engine code that
deals with promise reactions. Therefore, our implementation of `await`
will instead use a new variant of the `PromiseReactionHandler` enum.

To support this, however, there should be a way to invoke
PerformPromiseThen with the `on_fulfilled` and `on_rejected` arguments
being `PromiseReactionHandler`. To this end, `perform_promise_then` is
made closer to the spec, taking those arguments as `Value`s, and a new
`inner_promise_then` function is added.

This patch also refactors `PromiseReactionRecord` to store the
reaction type (whether it is a reaction to a promise fulfillment or
rejection), rather than storing it in `PromiseReactionHandler` only in
the empty case. This was not implemented this way initially, even
though it matches the spec better, because in the spec only empty
reaction handlers depend on the reaction type. However, this will no
longer be true if we add `await` continuations as a reaction handler.