refactor(ecmascript): Allow Symbol keys in function `Builtin`s
When specifying a built-in function with the `Builtin` trait,
previously there was no way to specify a function whose property key
is a symbol, since the property key is derived from the `NAME`
constant, whose type is `String`. It was possible (in fact, mandatory)
to specify a separate property key for getters, which could be a
symbol, but that was not the case for functions.

This resulted in function methods with a well-known symbol key having
to be handled manually in `create_intrinsic`, rather than having
`with_builtin_function_property` do the right thing. It also led to
such methods having the wrong `PropertyDescriptor` attributes, since
`with_property` forces you to set them manually; as well as to have
some methods with a well-known symbol key be mistakenly defined as
getters instead.

This patch fixes this by adding a `KEY` constant to the `Builtin`
trait, which allows specifying an optional `PropertyKey` to use
instead of the `NAME` constant. Since the `BuiltinGetter` trait (which
extends `Builtin`) already defined a non-optional `KEY` constant for
the same thing, this patch removes it, using the `Builtin` constant
instead.