feat(vm): Implement complex object binding patterns (#400)
This patch implements support for default values, nested destructuring
and computed rather than static property keys in object destructuring
bindings. It does so by building a slow path, separate from the
previous fast path of simple object binding, which implements the
entire destructuring in regular bytecode. This is parallel to the
distinction between simple and complex array binding introduced
in #394.