refactor(vm): Allow nested binding patterns to be simple (#401)
Array and object binding patterns can be either simple or complex,
where simple patterns are compiled in a way that jumps into a separate
execution context where only binding-related bytecode instructions are
supported.

Currently, if an array or object binding pattern contains another
binding pattern, that automatically makes the outer pattern complex.
However, that doesn't need to be the case, as long as none of the
nested patterns, no matter how deeply nested, have initializers or
computed property keys.

As it turns out, the spec has a syntax-directed operation called
`ContainsExpression` that checks exactly this, and which affects the
behavior of `FunctionDeclarationInstantiation`. Our implementation
currently does not use that, since we don't yet support bindings or
initializers in function parameters, which is why we don't yet
implement it. This patch adds an implementation for that, and uses it
to enable nested simple binding patterns.