feat(vm): Support the spread operator in function arguments
The call instructions (`DirectEvalCall`, `EvaluateCall` and
`EvaluateNew`) take their arguments from the VM stack, and get the
number of arguments from an immediate argument in the instruction.
This requires knowing the number of arguments at compile-time, which
does not work for the spread operator.

In order to support this operator while using the same instructions,
rather than adding new ones, the value `u16::MAX` is treated as a
magic constant indicating that the result value should be used as the
number of arguments.