feat(ecmascript): Implement the `Function` constructor
This patch implements the `Function`, `AsyncFunction` and
`GeneratorFunction` constructors, which parse the arguments as source
code and build a function from it.

It seems like this addition marks the first time that a function can
be constructed whose source text is not part of the ECMAScript code
context's source code. Therefore, we add an optional `SourceCode`
argument to `OrdinaryFunctionCreateParams` so the stringification
works properly.

One issue in the implementation of these functions is that the oxc
parser's API only exposes entrypoints to parse a whole script/module,
or an expression; while the spec for these constructors requires that
the function arguments and the function body be parsed separately, to
avoid injection.

Due to oxc not exposing this API, we can't follow the spec exactly
here. What we do instead is generate the spec-mandated source code for
the function (that would be returned when stringified), parse that as
a script, and throw if the AST doesn't match exactly one statement
which is a function declaration of the right type.