feat(engine): Implement interleaved garbage collection
This is a mostly-working interleaved, exact GC except that it is
currently built in a very unsafe way and exhibits undefined behaviour
in about 50 test262 cases. That's not good at all. I will need to
keep working on this to put it on a stronger footing, but the basic
idea is here.