feat(engine): Implement interleaved garbage collection (#443)
An interleaved, exact, stop-the-world GC implementation.

At present the GC is (mostly?) memory safe but not very reference-correct or useful. The collection is just "after every 256 instructions in the same bytecode" and the rest of the engine has no lifetime hints about when it's safe or unsafe to retain JS Values.