stash: try out usdt tracing
Linux perf does not understand the probes laid down by the usdt crate.
This mostly makes sense, as usdt is made by oxidecomputer folks and they
mainly target illumos and linux distributions with DTrace available.

My assumption is that we end up with the no-action register_probes() call
and hence we get no probe information. Not sure why that is but so it is.